requirements.txt
fastapi==0.88.0
uvicorn==0.20.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.10
pydantic==1.10.13
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pyotp==2.9.0 